Add support for log nodes in Beanstalk compiler
Summary:
The BM->BMG compiler now supports analyzing calls to math.log and torch.log, and lowering them to log nodes in the graph. Note that we do not yet support log nodes in the C++ code that implements the graph, but when we do, we will be ready.

Note also that the addition of `log(exp())` around the expression in `x(n)` does not change the DOT form of the graph because the graph builder detects that the expression inside is logically a constant and folds it away in the two samples.

Reviewed By: michaeltingley

Differential Revision: D18495064

fbshipit-source-id: df4fcb8d306f83bc917f45b0e0173937d7d0bde2