Refactoring of folder for pure functions
Summary:
I've gone with a table-driven solution for folding pure functions of one argument, so that we can easily add more as we see fit.

Thus far we've only got acos, exp and log from the math module, and exp and log from the tensor module, but I'll add more presently.

Reviewed By: michaeltingley

Differential Revision: D18516770

fbshipit-source-id: 2521662cb226a530d2f88a244b4a49c46837f3dd