Start implementing partially-constant expression optimization pass
Summary:
The constant folding pass deals with programs that have expressions that are entirely constant. This pass will deal with expressions that are partially constant and can be optimized accordingly.

To begin with, optimizations involving `and` where the leftmost operand is a constant.  `False and x` can be optimized to `False` regardless of whether `x` is a constant, and `True and x` can be optimized to `x` similarly.

In my next diff I'll do the same for `or`, and then similarly for conditional expressions and `if` statements.

Reviewed By: michaeltingley

Differential Revision: D18544505

fbshipit-source-id: 5fd099a8b2bfbb7d4545691f7e3376ccdc738cce