torch.neg is now constant folded and analyzed by beanstalk compiler
Summary:
The constant folder now treats torch.neg as a recognized pure function. If used on a non-constant, it is lowered into a NEG operation in BeanMachine Graph.

Note that the test case shows that NEG(NEG(x)) is NOT optimized to x.  I believe that it is a law that NEG(NEG(x)) == x, so I might make this optimization in a later diff.

Reviewed By: michaeltingley

Differential Revision: D18518302

fbshipit-source-id: 093aacc591e4ed90dce2c4a506f7e6d644965d25