Start implementing a cyclic-model detector in the transpiler
Summary:
We require that the graph produced when translating a BeanMachine model to BeanMachine Graph be acyclic, but of course it is easy to create a model that is cyclic.

Note that we must ensure that the graph is acyclic at the level of individual samples. That is, if we have a graph where `x(2)` depends on `y(1)` which depends on `x(0)`, that's fine; we do not want to say that `x(n)` is involved in a cyclic dependency.

The way we achieve this is, interestingly enough, by taking advantage of the fact that we compile first to a "lifted" program where all values are nodes. Those nodes are memoized, and all the `sample` methods are memoized, so if the memoizer ever detects a potential unbounded recursion where calling `x(2)` directly or indirectly calls `x(2)`, we can raise an error and halt the construction of the graph, instead of going into an unbounded recursion and crashing.

Reviewed By: michaeltingley

Differential Revision: D18665315

fbshipit-source-id: dd9383d59315b07c2845e0340a76f66056f4bf1c