Fix all_children so that it supports recursive list edits
Summary:
There was a known bug in my implementation of `all_children`. Suppose we have a recursive rewrite on a list that is returning a `ListEdit(edits)`.  The recursion should recurse on the elements of the list edits, possibly rewriting them further, but we were not doing so.

I've fixed that for `all_children` but not for the `some_children` or `one_child` combinators; I'll do those later.

Reviewed By: michaeltingley

Differential Revision: D18608792

fbshipit-source-id: 649b3e9704f18c02fb3c6aa670561c7f625bacff