Continue working on lowering of lists and loops
Summary: I've got some ideas of how to lower loops and lists more generally than the existing loop unroller for collections of constants. This will involve lowering both the loop itself and the list construction; in this diff I am just dealing with simplification of the loop to move the collection into a variable of its own, similarly to how we've done with return statements.  I'll work on lowering of list constructions in a later diff.

Reviewed By: michaeltingley

Differential Revision: D18820691

fbshipit-source-id: a4529d51b036ae37fdd4abd1e75afe288ae020bf