Continue to set up streamlined code generation
Summary: I want to change the code generation for the BM->BMG compiler so that (1) it is smarter about understanding arbitrary function calls, and (2) it keeps values that do not involve a sample operation as plain-old-Python-values.  I'm continuing to set up helper functions for that; this time, I've got the function, sample and Bernoulli handlers working together to support Bernoulli distributions.

Reviewed By: michaeltingley

Differential Revision: D19336479

fbshipit-source-id: 28f922ca6076b4e762760cc8647837a54f5b0f5a