Start trying to keep normal values normal.
Summary:
My initial idea in lowering a Python program into one that computes the graph was to wrap *every* value in a wrapper. Over the break I did some thinking about this and I now believe that this will just lead to a lot of unnecessary wrapping and unwrapping in non-trivial programs.

I've started to add helper methods to the graph builder that *avoid* adding a node to the graph if the node would consist solely of "normal" values; we only want to track nodes where one of the descendant nodes is a sample, typically.

The code generator does not yet use these; I'm going to add them in stages and test them independently, and then make modifications to the code generator.

Reviewed By: michaeltingley

Differential Revision: D19280702

fbshipit-source-id: 87e54f810e357c1d053ea0329a335b93b1e9ea9a