Add handlers for "dot" operator on samples
Summary:
I'm continuing my effort to add helper methods that will allow us to keep "non sample" values unwrapped when constructing the graph. The question I'm working on resolving with this diff is how to compile something like:

    y = x().exp()

where `x()` is a sample function that is logically a tensor, and the `exp()` is the instance function on tensors. We would like this to behave the same as

    y = torch.exp(x())

I will eventually lower the original code to something like:

    t1 = handle_call(x, [])    # will return a sample node
    t2 = handle_dot(t1, "exp") # will return a lambda
    t3 = handle_call(t2, [])   # will return an exp node with the sample as a child
    y = t3

Reviewed By: michaeltingley

Differential Revision: D19353074

fbshipit-source-id: 724b950b82511d6606791a80ac1848ca87ac61b9