Implement helper for adding functions to graph
Summary:
I want to make a generalized version of code which adds a function call to the graph, knowing just the function and its arguments. (Variadic parameters and named parameters are not yet supported; I will add those later.)

I will update the Beanstalk compiler to use this helper in a future diff; this diff is just putting it in place and unit testing it.

Reviewed By: michaeltingley

Differential Revision: D19321467

fbshipit-source-id: fceac3affc2fe4cf727de2255138c9480686242d