Improvements to unit tests for graph builder
Summary:
Recall that the idea of this compiler is that we transform the original program into a new program; when that new program is executed, the output is the graph.

I wish to make some changes to the original program transformation so that it is less aggressive about building up graph nodes; it should only produce graph nodes when necessary, and should otherwise keep values as "normal values".

I've created some helper functions to do so, which are not yet used; they will be used in subsequent diffs.  However, I've created the unit tests for those methods now, so that I know where they do and do not work.

This large diff adds all the unit tests for those new methods.

Reviewed By: michaeltingley

Differential Revision: D19385201

fbshipit-source-id: 4e3bd59d2807f148c2bc4745cda0e43a72f18701