Start implementing a type system on graph nodes
Summary: I've started implementing a simple type system on graph nodes; this is just a placeholder until we figure out the real type system. In particular, we will need to figure out which node topologies have no types because they are illegal, like "bernoulli plus bool", or something like that.

Reviewed By: michaeltingley

Differential Revision: D19526178

fbshipit-source-id: b5d808739f9a2a6eaaebf4b9ce437030c7fe6c23