BMG compiler updated to use handle functions
Summary:
First off, recall again the design of this compiler. We lower the original BM program into a version of it which, when executed, produces a graph builder when executed. We can then execute that graph, or serialize it in a variety of formats, including an optimized program that produces the graph again.

In my original prototype BM->BMG compiler, I was trying to lift all runtime data into a graph node, and then discard what we don't need. However, this means we are potentially doing a lot of computations on "normal" values that are ultimately irrelevant because they will be discarded from the graph.

The new idea is to keep all "normal" values as normal values. The only operations we lift into the graph algebra are those that involve samples somehow: adding two samples together, say.

The "handle" methods on the graph builder take care of these semantics: if the result of the operation is a regular value, it is returned; if it is a graph node, that is returned.

This checkin gets the basic functionality working; however, there is still a long way to go:

* A great many features are missing from the implementation of BMG; I've turned off the tests for lowering into BMG until we fix that. See my "missing features of BMG" document for details.
* Lists are not handled correctly
* Function calls involving named parameters and variadics are not handled correctly
* Function calls where the function is in the model and takes a sample are not handled correctly.
* And so on.

I'll start adding these features in subsequent diffs.

Reviewed By: michaeltingley

Differential Revision: D19566832

fbshipit-source-id: 48f038bb8524bbdca36c821706977344dc0e5104