Add normal distribution to graph builder
Summary:
I've added support for the normal distribution to the graph builder; we have not yet added this to the C++ implementation of BMG but will eventually.

My purpose here is simply to get support for a couple more common distributions so that I can make more realistic test cases that implement real models.

In this diff for example I've added a test case that implements the Neal's Funnel model, a simple model involving two normal distributions.

I'll add support for a uniform distribution in an upcoming diff.

Reviewed By: michaeltingley

Differential Revision: D19846417

fbshipit-source-id: 208930371e3816346010e2c6b488de8d7ea9a8d3