Implement correct lowering for calls to ordinary functions in module
Summary:
What happens when we have a lowered-to-BMG-producer BM module where a function in it takes as an argument a graph node, where that function is just an ordinary function, not a sample?  What should happen is we should just pass the graph node in and process it like any other data.

To achieve this we'll need to know that the function in question was built to handle graph nodes. We achieve this by adding a note to the generated module, and then checking at runtime whether the function was from such a module.

Getting graph nodes passed to sample functions will be considerably trickier and will require a new node type ("index") added to the graph builder, plus the ability to determine the support of a graph node. I'll be adding those features in upcoming diffs.

Reviewed By: michaeltingley

Differential Revision: D19670382

fbshipit-source-id: 5650847a9b5e190950c992c0b75b982eff1994b8