Add Beta distribution to graph builder
Summary:
I've added support for the Beta distribution to the graph builder; we'll add support to BMG proper at a later date. This will enable me to build more realistic models for testing the compiler.

The test case is very simple: get a beta-distributed "fairness" from a mint, have the mint make a coin with that fairness, and flip it:

{F228431139}

Reviewed By: michaeltingley

Differential Revision: D19868248

fbshipit-source-id: 1534e96a7351f26ddc6820969b98838cbe98f9f0