Implement matrix multiplication support in compiler
Summary:
We'll need to support both normal tensor multiplication and matrix multiplication in BMG. To that end, I've added support for generating matrix multiplication nodes in the graph builder; any use of torch.mm or torch.Tensor.mm functions will generate such a node.  As with other nodes, I'll automatically constant-fold operations on constants.

I have not yet added support in the C++ implementation of BMG; that will come in a later diff.

The test cases so far are just unit tests. In an upcoming diff I'll add a test with a realistic model and show that it compiles to the correct graph.

Reviewed By: michaeltingley

Differential Revision: D20015546

fbshipit-source-id: f45b2b2d544aee15cf4ae176f51fa3d569dabfb3