Fix add_map in builder so that it can be memoized
Summary:
add_map in the graph builder took a list, but that makes it ineligible for memoization because lists are not hashable in Python. I've made it variadic instead, which both looks nicer at the call site, and allows memoization.

I also improved the test case to be more realistic, and added a second call to demonstrate that the memoization makes it idempotent.

Reviewed By: michaeltingley

Differential Revision: D19821703

fbshipit-source-id: 6a05ba3d44523e6669da148bbd237821579a68d1