More work on helpers for lifting results of instance methods into graph
Summary:
My work on getting instance method invocations such as the "add" instance method of a tensor into the graph continues.

In this diff I've gotten additions of this form working:

    t = tensor(1.0)
    a = t.add
    s = x() # a sample of tensor type
    r = a(s)

This should make `r` represent an ADD node with the tensor 1.0 as one child and the SAMPLE node as the other.

Again, the reason to make this work is that  in a coming diff I will transform

    r = tensor(1.0).add(x())

into the form above in a simplifying pass, and so we will need to be able to compile such a program into a graph.

The helper functions will be complete in the next few diffs and then I will make the compiler use them.

Reviewed By: michaeltingley

Differential Revision: D19549269

fbshipit-source-id: 368715f83dae93601b720d6ff0fb46984073036c