Implement "logits" support for Bernoulli distributions in BM->BMG compiler
Summary:
We now correctly handle distinguishing between "probs" and "logits" flavours of the Bernoulli distribution.

The details can be summed up by this model and corresponding graph:

	sample
	def toss():
	    return Bernoulli(probs=0.5)
	sample
	def toss2():
	    return Bernoulli(logits=0.0)
	sample
	def toss3():
	    return Bernoulli(probs=toss())
	sample
	def toss4():
	    return Bernoulli(logits=toss())

logits=0.0 and probs=0.5 both represent "fair coin", so toss and toss2 are logically the same distribution. Notice that we identify this fact and represent it in the graph.

However, the meaning of `Bernoulli(probs=toss())` is "mint either a double-headed or double-tailed coin", and the meaning of `Bernoulli(logits=toss())` is "mint either a fair coin or a coin with probability e/(e+1)". We must therefore notate that in the graph.

We have not yet added support to BMG proper for this new kind of Bernoulli distribution.

Reviewed By: michaeltingley

Differential Revision: D20007663

fbshipit-source-id: 808e580c4d740738c5fcd08575b01712a9ed7109