Implement nodes for index/map operations
Summary:
In order to represent probabilistic workflows we will need to be able to generate "choice" nodes, where a probabilistic workflow is chosen probabilistically.  We have not added these nodes to the C++ implementation of BMG yet, but to test the compiler I've added a prototype of these nodes to my builder.

Adding gear to generate these nodes from BM source code will come in a later diff.

In my test case I've implemented the workflow "flip a fair coin; if you get 0.0, flip another coin; if you get 1.0, flip a coin and add 2.0".  I've verified that the support of the indexed distribution is 0, 1, 2, 3.

The graph looks like this; the convention I've implemented here is that "even number" nodes are keys and "odd number" nodes are values but we can change that up later if we prefer a different convention.

{F228005253}

Reviewed By: michaeltingley

Differential Revision: D19781386

fbshipit-source-id: 81ddc131043b26388f6b9081b43263e457563879