Implement code that computes support of nodes in a graph
Summary:
Many of the nodes in a graph represent a discrete distribution and therefore have a finite "support" -- that is, the finite set of values that can be sampled.

To correctly generate graphs for certain kinds of models will need to be able to compute the support of a graph node to add a new node to the graph. For example, the support of a Bernoulli would be 0, 1, and the support of the sum of two Bernoullis would be 0, 1, 2.

This diff adds code to compute the support of graph nodes. However, it makes certain simplifying assumptions that will need to be corrected later; in particular it assumes that all Bernoulli nodes are one-dimensional. Torch supports `Bernoulli(tensor([0.5, 0.5]))`, say, representing "flip two coins", and so the support of this should be the four possible outcome tensors. I'll fix that in a later diff.

Unfortunately I discovered that tensors do not play nicely with sets; in Python `{tensor(1.0)} == {tensor(1.0)}` is false, and worse, `{tensor(1.0), tensor(1.0)}` is not deduplicated! I've created my own list-based set that only may contain tensors. Though this is linear in its operations instead of O(1), the sets will be small.  Again, we can make this more sophisticated and performant in the future.

Reviewed By: michaeltingley

Differential Revision: D19732228

fbshipit-source-id: 235ecd04f9c5011e1fc6838987ed5a5a81096148