Create function transformation for probabilistic workflows
Summary:
When lifting a sample method in BM to BMG, it returns a specific distinct sample node for each distinct combination of arguments.

If the arguments are just ordinary values, memoization suffices to implement these semantics. But what happens when we call a sample method and the argument is itself a probabilistic workflow?

In this case we generate an indexing operation; we call the sample method once for each possible value in the support of the probabilistic workflow, and build a map from those possible values to the sample operations that are the outcomes.

This decorator implements this function transformation.

I have not yet modified the compiler to generate the decoration; this just implements and tests it.  I'll modify the compiler in a later diff.

The test cases illustrate some scenarios.

The first is simple: we flip a 50% coin to get either 0.0 or 1.0. Based on the outcome of that coin, we flip one of two different coins; either a 25% coin or a 75% coin:

{F228088204}

As noted in the previous diff in this stack, the "map" uses the convention that even numbered edges are keys and odd numbered edges are values.

The second is rather more complicated, but if you look at the graph carefully it makes sense.  We are flipping two coins. The first, x, we flip to get 0.0 or 1.0. The second, y we flip and add 2 to get 2.0 or 3.0.  That gives us four possible options for (x, y); those four possibilities correspond to four samples, but two of the samples are in this case from the same distribution. Notice that the graph is deduplicated, but the key point is that there are SIX sample operations: One for the first argument, one for the second argument, and then four samples for the four possible control flows.

{F228088302}

Reviewed By: michaeltingley

Differential Revision: D19797959

fbshipit-source-id: 285454ee64ba222f4f5b13e5025cf034d3cbe1c3