Start implementing support for named arguments on known functions
Summary:
My changes to support calls with named arguments continues. In this diff we see that we now support the `input` named argument on calls to `torch.exp` and `torch.log` **regardless of whether the argument is a normal value or a graph node**.

I've started making the dispatch here table-driven rather than a huge if-then statement as well.

In upcoming diffs I'll continue to make these changes until all built-in functions support named arguments.

Reviewed By: michaeltingley

Differential Revision: D19960012

fbshipit-source-id: be095af29a5757956ead791d6e8bc93cdb33b75d