Implement support in graph builder for simple calls with keyword args
Summary:
Another in my series of diffs for adding support to calls with keyword args to the compiler.

In this diff, the graph builder now operates correctly in cases where it does *not* have to build a graph node:

* Any call to a method that is already in a module that has been lowered -- these functions know how to handle the arguments if they are graph nodes, so they can just be called
* Any call to a normal method with normal arguments

The test case here shows that when we call `Bernoulli` with a named argument and the argument is just a normal value, we just make the call. (The distribution object is lowered to a graph node when it is sampled.)

In upcoming diffs I'll work on calls where the keyword argument is a graph node; for example `torch.exp(input=x())` where `x()` is a sample node.

Reviewed By: michaeltingley

Differential Revision: D19917699

fbshipit-source-id: bbd44dd02f1837dec01213843268780c444e5b87