Test case for Bayesian logistic regression
Summary:
The idea of this model is that we have a tensor X that has N rows and K+1 columns; the leftmost column is 1.0.  We also have -- not shown in this test case -- a tensor Y with N rows that are either true or false; these are observations that a predicate has classified the corresponding rows of X.

We wish to know the parameters beta() of the plane which divides the true from false values. We model those parameters as normal values centered on zero with assumptions about the approximate magnitude of the parameters as the standard deviation. Each observation is then modeled as a coin toss that classifies each row as true or false:

    K = 2
    # This is just arbitrary test data
    # In a realistic case N would be much bigger.
    X = tensor([[1.0, 10, 20], [1.0, -100, -190], [1.0, -101, -192]])
    intercept_scale = 0.9
    coef_scale = [1.2, 2.3]
    sample
    def beta():
        return Normal(zeros((K + 1, 1)), tensor([intercept_scale] + coef_scale).view(K + 1, 1))
    sample
    def y():
        return Bernoulli(logits=X.mm(beta()))

This model has many features not present in earlier "real world" test cases:

* Bernoulli uses logits, not straight probabilities
* We have a matrix multiplication where one of the operands is a sample
* We have multidimensional tensors, tuples and lists; though none of them involve samples, it's good to demonstrate that we can support them.

The graph generated by the compiler is:

{F229131770}

Reviewed By: michaeltingley

Differential Revision: D20042388

fbshipit-source-id: ff1422e89b68e2f5700784f2e13edbc719368a02