Add support to handlers for instance methods on graph nodes
Summary:
I wish the Beanstalk compiler to have the property that the compiler will produce the same graph for these two obviously equivalent programs:

    s = x() # a sample
    r = s.exp()

and

    s = x() # a sample
    e = s.exp
    r = e()

In both cases, r must logically be an EXP node pointing to a SAMPLE node in the graph.

The reason I particularly want this is because it will be convenient for me to transform the first program into the second as part of the compilation, so I need to get it working in the second case.

This diff gets the helper functions that I will be using to achieve this working for exp, log, neg and logical_not instance functions on tensors.

I'll work on cases like

    s1 = x(1)
    s2 = x(2)
    a = s1.add
    r = a(s2)

in a later diff.

Reviewed By: michaeltingley

Differential Revision: D19544513

fbshipit-source-id: ac564048c7114c8920a1e3c6dbec72527b79eb1f