Continue to add support for named arguments in known-function calls
Summary:
Named arguments are now supported for all the "static" `torch.foo` operations and calls to the Normal and Beta distribution constructors.

The `math.log` and `math.exp` calls do not support named arguments in the first place, so I've just added those to the table; they'll keep on working as normal.

Support for named arguments in calls to tensor instance functions will come soon. Also I have not done Bernoulli yet because it has a small complication to deal with; it supports both `probs` and `logits` named arguments but not both. I'll have to write special code for that.

Reviewed By: michaeltingley

Differential Revision: D19962424

fbshipit-source-id: f16470ddb74a6bc6a633ca8a8a8c8f872783cd9d