Add test that demonstrates probabilistic workflow
Summary:
I've added a test that demonstrates how probabilistic workflows compose in the BM->BMG compiler.

What we have here is a model where we flip three coins. First we flip a fair coin.  Based on that outcome we generate one of two unfair coins and flip it.  Based on *that* outcome, we generate either a double-headed or double-tailed coin.

The result is a graph which illustrates this workflow by showing all the possible paths through the program.

{F228302924}

Once again, recall that the convention we're going with for now is "even indices are keys, odd indices are values". So if we get 0.0 (index 0) then we choose sample Bern(0.25) (index 1), and so on.

Reviewed By: michaeltingley

Differential Revision: D19838304

fbshipit-source-id: 34b749ffdfa48ea3ddb398c50d37ae5b70a8ed60