Add keyword argument support to single-assignment rewriter
Summary:
My simple single assignment rewriter now supports calls with keyed arguments, like `x = g(123, y=456)`, rewriting them to `t1 = 123, t2 = 456, x = g(t1, y=t2)`.

In a later rewriting pass I'll then transform this into something like

    x = bmg.handle_function(g, [t1] {y=t2})

so that the function call can be analyzed to see if it is one we know about and need to generate a graph node for.

That will come in a future diff.

Reviewed By: michaeltingley

Differential Revision: D19913082

fbshipit-source-id: e726487d95b2caba102460da19812eb67ef9ebbc