Start adding support for calls with keywords to compiler
Summary:
We're going to need to be able to handle situations like `Bernoulli(probs=0.5)` that we do not handle right now; this will have to compile to a node the same way that `Bernoulli(0.5)` does. And of course we will also need to distinguish between `Bernoulli(probs=x())` and `Bernoulli(logits=x())`.

These changes are not huge in themselves but they involve changing a lot of test cases, so I'm going to approach it a bit at a time. In this diff I just change `handle_function` so that it can accept a keyword dictionary which it ignores, and I've changed the code which generates calls to `handle_function` so that it always passes an empty dictionary.

In upcoming diffs I'll implement the compiler to generate the appropriate dictionary, and implement `handle_function` to handle it appropriately.

I'll also eventually add support for calls with `*` and `**` arguments.

Reviewed By: michaeltingley

Differential Revision: D19913093

fbshipit-source-id: 7612809162ace621c2f05f005f07ab5d8fd98bc4