Compiler support for function calls with keyword args
Summary:
This is the second diff in my series where I'm gradually adding support to the compiler for function calls with keyword arguments.

In this diff, we generate a correct dictionary given a collection of keyword arguments.

The code generation is a little odd; for

    f(x, a=b, c=d)

I do NOT generate

    handle_function(f, [x], {'a'=b, 'c'=d})

But rather

    handle_function(f, [x], { **{'a'=b}, **{'c'=}})

The reason for this oddity is: in an upcoming diff I'm going to support

    f(x, a=b, **kwargs, c=d)

and clearly it will be easy to splice that in there correctly in this format.

Note that the handle_function function does not correctly handle this new input yet! I'll add that support in a future diff. This diff just gets the compiler working.

Reviewed By: michaeltingley

Differential Revision: D19912996

fbshipit-source-id: 474dc0299321446969dfd04bc3cdfd628e40df28