Finish implementing basic support for indexing operation in BM->BMG compiler
Summary:
The Beanstalk compiler now generates graphs for indexing operations of the form `collection[index]` where the collection is a list and the index is a graph node. We generate a map/index pair of nodes in the graph.

As noted previously in this diff stack, the BMG implementation of these nodes is not yet implemented; we'll do that in a future diff.

The test case added is a simplified version of a model for detecting fake accounts based on what accounts they send friend requests to, and what accounts accept those friend requests:

    FAKE_PRIOR = 0.001
    # One entry per user
    FAKE_REQ_PROB = tensor([0.01, 0.02, 0.03])
    REAL_REQ_PROB = tensor([0.04, 0.05, 0.06])
    REQ_PROB = [REAL_REQ_PROB, FAKE_REQ_PROB]
    REAL_ACC_PROB = tensor([0.99, 0.50, 0.07])
    sample
    def is_fake(account):
      return Bernoulli(FAKE_PRIOR)
    sample
    def all_requests_sent(account):
      return Bernoulli(REQ_PROB[is_fake(account)])
    sample
    def all_requests_accepted(account):
      return Bernoulli(REAL_ACC_PROB * all_requests_sent(account))

This generates the graph:

{F229812806}

Reviewed By: michaeltingley

Differential Revision: D20129074

fbshipit-source-id: 95384eba51978fc19a1c2ccf65b9f29acc708d2d