Add support to BM->BMG compiler for uniform distributions
Summary:
I've added support to the Beanstalk compiler for uniform distributions; the node for this is just in my graph builder, not in the C++ implementation of BMG. We'll add it to BMG later.

This allows me to write a test case for another real-world-inspired model: linear regression. (Thanks to Kinjal for the model.)

    sample
    def theta_0():
	    return Normal(0,1)
	sample
	def theta_1():
	    return Normal(0,1)
	sample
	def error():
	    return Uniform(0,1)
	sample
	def x(i):
	    return Normal(0,1)
	sample
	def y(i):
	    return Normal(theta_0() + theta_1() * x(i), error())

Here we see the generated graph for y(0), y(1), y(2). It is a little hard to read because of all the deduplication; can you identify the meaning of each of the nine sample nodes? :)

{F230007965}

Reviewed By: michaeltingley

Differential Revision: D20151789

fbshipit-source-id: ecc696e00441ac127c154cd8aa03f0ba5a3acdd5