Start tracking tensor shape information in the BMG graph builder
Summary:
My BMG graph builder now deduces information about what the shape of the tensor is for each node. For example, if we have a `Bernoulli(tensor([0.5, 0.5]))` sample then its shape is `torch.Size([2])`, indicating a one-dimensional vector of width 2.

The reason we need this is because we currently do not correctly compute the support of distributions, which is necessary for correctly generating graphs where one sample is an index to another sample.

This will be particularly important when we add support for categorical distributions because the support of the distribution is a function of the shape. That is, if we have `Categorical(tensor([0.4, 0.3, 0.2, 0.1]))` then we will need to know that the support is integers `0` through `3`, computed from the width of the input tensor.;

The bugs in support computation are not fixed yet. That will come in an upcoming diff.

Reviewed By: michaeltingley

Differential Revision: D20171129

fbshipit-source-id: 9af519d9aa40f0fa51385478478bd193ecc09c97