Implement test case for Gaussian mixture model
Summary:
The Beanstalk compiler is now capable of compiling a simple Gaussian mixture model into a graph.

The idea here is that we have, say, three means. To sample from the mixture we first choose a mean with a categorical distribution that generates 0, 1 or 2 in some ratio, then determine the mean by sampling from a normal distribution centered on zero, and then sample from a normal distribution centered on the given mean.

See the source code below for the model. A slightly simplified version of the generated graph is shown below; here we are generating one sample from the mixture model.

{F231293834}

Reviewed By: michaeltingley

Differential Revision: D20400787

fbshipit-source-id: 4b0ae616a34aeceb164d1235df8d2a88288cdbde