Implement support computation for Bernoulli distributions
Summary:
To construct a graph where we pass a sampled value as the "index" to another sample function, we need to know the support of the sampled value's distribution.

For Bernoulli distributions we had it hard-coded to 0.0 and 1.0, but that's not right in general; we might be flipping two coins via `Bernoulli(tensor([0.5, 0.5]))` for instance.
The graph generator now correctly generates the support of an arbitrary Bernoulli distribution.

(Note that we cannot simply use the `enumerate_support` method of the `Bernoulli` object since it does not actually enumerate the support. It enumerates the first and last members of the support.)

Here for example we flip two coins, take their average, and then use that to choose whether to flip a fair, double-headed or double-tailed coin:

    sample
    def x(n):
      return Bernoulli(n.sum()*0.5)
    sample
    def y():
      return Bernoulli(tensor([0.5,0.5]))
    sample
    def z():
      return Bernoulli(x(y()))

This generates this graph. (Remember that we are using the convention for maps that even numbered edges are "keys" and odds are the corresponding values.)

{F230346376}

Reviewed By: michaeltingley

Differential Revision: D20205397

fbshipit-source-id: 6acbdaf753bbe5ccbfe296e3fac525a69f46a786