Add support for categorical distribution in Beanstalk compiler
Summary:
The BM->BMG graph builder now supports the Categorical distribution. When a probabilistic workflow uses a sample from a categorical to call an "indexed" sample method, we correctly add a branch for each member of the support to the generated graph.

For example, consider this toy model. We choose a number `y()` from 0 to 2 in the ratio 2:6:8. Based on that we generate three different samples: `x(0)` is "flip a fair coin", `x(1)` is "choose a number from 0 to 2 in the ratio 1:3:4" and `x(2)` is "flip an unfair coin".

That then gives us a number from 0 to 2, which we use to mint coin `z()` and sample it:

```
sample
def x(n):
  if n == 0:
    return Bernoulli(0.5)
  if n == 1:
    return Categorical(tensor([1.0, 3.0, 4.0]))
  return Bernoulli(0.75)

sample
def y():
  return Categorical(tensor([2.0, 6.0, 8.0]))

sample
def z():
  p = x(y()) * 0.25
  return Bernoulli(p)
```
The graph looks like this -- recall again that this graph uses the somewhat unfortunate convention that even numbered edges coming out of the map are "keys" and the corresponding odd numbers are the associated value:

{F231274528}

A fun thing to notice here is that the graph builder determines that a categorical with the ratios 2:6:8 and a categorical with the ratios 1:3:4 are the same distribution, and it deduplicates them. We get this for free because of the way PyTorch builds distribution objects and normalizes their parameters in the process.

We have not yet added code in the C++ graph library to support these; when we do, I will update the code to generate those nodes and test it.

I've implemented support here for categorical distributions where the tensor describing the likelihood of each element is log-odds, but I have not yet tested that code path. I'll do that in a later diff.

Differential Revision: D20234769

fbshipit-source-id: 674bc4894d5e318474e52e7c214ac20a36fcb879