Integrating testing of "unassigned expressions" rewrites into main ssa testing
Summary:
The purpose of this diff is to add the newly added rewrite for
"unassigned expressions" to the default set of rewrites that make upp ssa.
This way, past and future tests would automatically take into account this
transformation.

This is now done. It did require some changes to the old tests, and in particular, to use the trick of naming some previously unassigned expressions to avoid triggering that transformation and other transformations so as to avoid seeing global changes to the indices used in the fresh names.

Reviewed By: ericlippert

Differential Revision: D20519498

fbshipit-source-id: 772ba7a6c116bb99ca702e8805df75c7e6497577