Create test case that shows compiler can handle CLARA model
Summary:
I've created a simplified CLARA model as a test case.

We have three kinds of documents: photos of sheep, goats and cats.

We have two classifiers, Alice and Bob.

Our prior on the distribution of sheep, goats and cats is uniform; we do not know anything at all about the distribution. Any categorical is as likely as any other so far as we know.

Our prior on classifier behaviour is not flat; we assume that if you're given a photo of a sheep you might think it is a goat, but not a cat, and so on.

We then observe our two classifiers making a classification of two documents, and compile the program into a graph. (I've simplified the maps slightly to make them easier to read.)

That results in four samples from y() which kicks the whole thing off:

{F232015572}

Many thanks to Naz and Narges who helped me understand some of the details of this model.

Differential Revision: D20568074

fbshipit-source-id: b679f1991e6e0d0a8649aae2b8594a288785704b