Stylistic changes to _transform_with_name
Summary:
Mostly renaming and in one case tightening a type annotation a bit.
The type annotation is to replace ast.AST to ast.Expr in the return type of
one of the two functions passed to _transform_with_name. The renmaing is to
use names such as extract_expr and build_new_term for the arguments to this
function and new_name and lhs for the arguments of the functions that are
typically passed to this function.

Reviewed By: ericlippert

Differential Revision: D20657278

fbshipit-source-id: ef8faab84b13eba506701ccd56d98b31b97a5bcf