Rewrite for while loops
Summary: We rewrite while looks to moving condition expression inside the loop and else clause to follow it. There is also a trivial case when the condition of the loop is already True where no work needs to be done, unless there is an else branch. That else branch is dead code, and should be removed. This initial diff only implements that very last part.

Reviewed By: ericlippert

Differential Revision: D20672034

fbshipit-source-id: 93866cec5cdd4bc6156baa7f6a17bb14fea3c261