Rewrite rule for while (3/3)
Summary: This diff adds a single rewrite for while that works by applying either of the two rewrites. Caveat: The current diff exposed the fact that the two rewrites are not yet entirely compatible, and there is a test marked TODO that goes into an infinite loop right now as a result.

Reviewed By: ericlippert

Differential Revision: D20677297

fbshipit-source-id: 249d3de217a4b1b84ad19f7224677580a965157e